Tested reaching SunriseSunset API with GlobalScope.launch

+++ app/src/main/java/com/elliott/catcope/data/network/SunriseSunsetNetworkDataSource.kt
@@ -8,7 +8,7 @@ interface SunriseSunsetNetworkDataSource {
     val downloadedSolarEvents : LiveData<SunriseSunsetResponse>
 
     suspend fun fetchSolarEvents(
-        latitude: Float,
-        longitude: Float
+        latitude: Double,
+        longitude: Double
     )
 }
