+++ app/src/main/java/com/elliott/catcope/data/network/SunriseSunsetNetworkDataSourceImpl.kt
@@ -7,16 +7,15 @@ import com.elliott.catcope.data.db.entity.SolarEventsEntry
 import com.elliott.catcope.data.response.SunriseSunsetResponse
 import com.elliott.catcope.internal.NoConnectivityException
 
-class SunriseSunsetNetworkDataSourceImpl(
-    private val sunriseSunsetApiService: SunriseSunsetApiService
-) : SunriseSunsetNetworkDataSource {
+class SunriseSunsetNetworkDataSourceImpl(private val sunriseSunsetApiService: SunriseSunsetApiService)
+    : SunriseSunsetNetworkDataSource {
 
     private val _downloadedSolarEvents = MutableLiveData<SunriseSunsetResponse>()
 
     override val downloadedSolarEvents: LiveData<SunriseSunsetResponse>
         get() = _downloadedSolarEvents
 
-    override suspend fun fetchSolarEvents(latitude: Float, longitude: Float) {
+    override suspend fun fetchSolarEvents(latitude: Double, longitude: Double) {
         try {
             val fetchSolarEvents = sunriseSunsetApiService
                 .getSolarEventTimes(latitude, longitude)
@@ -26,4 +25,7 @@ class SunriseSunsetNetworkDataSourceImpl(
             Log.e("Connectivity", "No internet connection", e)
         }
     }
+
+    
+
 }
